quick commit

+++ localsuppress2014/clustering.py
@@ -114,8 +114,8 @@ def plot_xi(z):
             ax.set_ylabel(r"$\xi(r)$")
             ax.set_yscale("log")
             ax.set_xscale("log")
-            print "saving fig","halo_"+str(long(abs(mag)))+"_xi"+str(z)+".pdf"
-            fig.savefig("halo_"+str(long(abs(mag)))+"_xi"+str(z)+".pdf",bbox_inches='tight',pad_inches=0)
+            print "saving fig","halo_"+str(abs(mag))+"_xi"+str(z)+".pdf"
+            fig.savefig("halo_"+str(abs(mag))+"_xi"+str(z)+".pdf",bbox_inches='tight',pad_inches=0)
             print "done"
 def main():
     plot_xi("6.00")
